pyproject.toml
[tool.poetry]
name = "fastapi-deploy-sample"
version = "0.1.0"
description = ""
authors = ["himantyu-yuma"]
readme = "README.md"
packages = [{include = "fastapi_deploy_sample"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
websockets = "^11.0.3"
deta = "^1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.4.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
